backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pydantic